fix(geocoder): no more new geocoder load on render and refactor
